Fixed getAPM.
git-svn-id: https://bwapi.googlecode.com/svn/trunk@3395 273ecf2c-f24b-0410-b6a7-e9298bb4873f

+++ bwapi/BWAPIClient/Source/GameImpl.cpp
@@ -1019,7 +1019,7 @@ namespace BWAPI
   {
     return data->instanceID;
   }
-  double GameImpl::getAPM(bool includeSelects)
+  int GameImpl::getAPM(bool includeSelects)
   {
     if ( includeSelects )
       return data->botAPM_selects;
