+++ bwapi/include/BWAPI/Game.h
@@ -408,7 +408,7 @@ namespace BWAPI
       virtual int  getInstanceNumber() = 0;
 
       /** Retrieves the bot's APM. Can include or exclude select commands. */
-      virtual double getAPM(bool includeSelects = false) = 0;
+      virtual int getAPM(bool includeSelects = false) = 0;
 
       /** Changes the map to the one specified. Changes do not take effect unless the game is restarted. */
       virtual bool setMap(const char *mapFileName) = 0;
